requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.7.27.1
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.1
coreschema==0.0.4
cryptography==3.1.1
cssselect==1.0.1
decorator==4.1.2
Django==1.11.17
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-rest-swagger==2.2.0
djangorestframework==3.6.4
djangorestframework-sav==0.1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipython==6.1.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
lxml==4.5.2
Markdown==2.6.9
MarkupSafe==1.1.1
olefile==0.44
openapi-codec==1.3.2
parsel==1.2.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==7.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==17.3.0
pytz==2017.2
queuelib==1.4.2
requests==2.18.4
Scrapy==1.4.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.17.2
six==1.10.0
traitlets==4.3.2
Twisted==17.5.0
uritemplate==3.0.0
urllib3==1.22
w3lib==1.18.0
wcwidth==0.1.7
zope.interface==5.1.0